requirements.txt
pandas
requests
bs4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv